Port latest cache create behavior
In az cli 2.70 we introduced an improvement for `az acr cache create`,
where this command does not need to issue a GET registry request first.

This change is needed for users with cache rule permission only. If a user
only has access to cache rule resource, the GET registry request would
fail with 503.

This commit ported this change in `arccache` extension, so that even
after this extension is installed, user can still run `acr acr cache
create` command.