fix(cssc): 31694510, 31694600 Check for existence of cssc tasks before calling list command + enhance log message (#32)
This PR adds a check to ensure that the cssc tasks exist before calling
the list command. It also displays that the list command is executed for
the last n days. Updated tests as well.

With this, below 2 bugs are addressed:
1.
https://msazure.visualstudio.com/AzureContainerRegistry/_workitems/edit/31694510/?view=edit
Before Fix: 

![image](https://github.com/user-attachments/assets/61ced20f-d8c4-42d6-bc8d-67e3ae299ed1)
After Fix:

![image](https://github.com/user-attachments/assets/807745a3-71f5-471c-89b5-62161fbd3fd1)

2.
https://msazure.visualstudio.com/AzureContainerRegistry/_workitems/edit/31694600/?view=edit
After Fix: added a line to indicate list executed for last n days:

![image](https://github.com/user-attachments/assets/fe832467-d337-4f18-a2c7-f713452e5c3a)

---------

Co-authored-by: Ruchi Maheshwari <rumahe@microsoft.com>