fix(cssc): 31646926 fix 'list' command filtering by '--run-status' (#40)
Fix filtering on the `list` command by `--run-status` to make it correct
once again.
Now the `--run-status` will filter the image status depending on the
final status of both the scan and patch instead of only one of them.
Meaning that for an image to be marked as `FAILED` scan or patch have to
be `FAILED`, for an image to be marked as `SUCCEEDED` both scan and
patch have to be `SUCCEEDED` or `SKIPPED` as that is treated as a
successful execution, and for an image to be marked as `SKIPPED` the
scan has to be `SUCCEEDED` and the patch has to be `SKIPPED`.

This PR also includes a unit test to catch issues with the filtering.

In addition, includes fixes for unit tests that are run without azure
credentials, adding the required mocks to avoid that dependency.